feat(get-how-many-duplicates): add solution
This a solution for a interview challenge that I failed, so after
thinking a little about it the day after I could realize a nice
approach for the challenge.
The question was "many `str1`s can be constructured of the `str2`?".
Here is the solution!
I wrote the code in both TS and Dart languages for practice purpose.